Update patch set 1
Patch Set 1:

I believe the intention of this patch is to only modify the behaviour when GLOBAL_MOTION is enabled.  At the moment there are test failures and coding rate loss when GLOBAL_MOTION is turned on.  

Do you think this patch changes the behaviour when GLOBAL_MOTION is disabled?

Patch-set: 1