Update patch set 1
Patch Set 1:

I take back my previous comment.  This patch _is_ expected to change the behaviour both with and without global motion.

The intention of the patch was to allow a variety of bdrate experiments to be performed, after which this patch will be abandoned or tidied up.

Patch-set: 1
Label: Code-Review=0