Update patch set 7
Patch Set 7:

(1 comment)

Yes, these are just test to make these wrappers easier to maintain.  And the purpose of the wrappers is to make simd optimised code easier to write and maintain.  They're used in the CLPF code and eliminate the need for a lot of different versions having almost identical code and the maintenance nightmare of updating code for SSE2, SSSE3, SSE4.1, AVX2 and NEON for every change.  I'm not putting any pressure on others to use them, though.

Patch-set: 7
Label: Code-Review=0