Improve rdopt decisions for ext-inter
When predicting interintra modes, the previous code did not provide
the intra predictor with the correct context during rdopt. This
patch adds an explicit 'ctx' parameter to the relevant functions,
to provide this context.

In addition, interintra prediction for rectangular partitions was
overwriting the inter prediction it generated with part of its
intra prediction, leading to suboptimal decisions. Disable interintra
modes on rectangular partitions for now, until we have rectangular
intra predictors.

Change-Id: I7458c036c7f94df9ab1ba0c7efa79aeaa7e17118