Enable the activity masking of PVQ
However, the distortion function of Daala should be also added
to complete the activity masking working.

Note that PVQ QM matrix (i.e. scaler for each band of
transform block) is calculated in decoder size as exactly same
way in encoder. In Daala, this matrix is written to bitstream
and decoder does not generate it.

Change-Id: I44bfb905cb4e0cad6aa830a4c355cd760a993ffe