Enable the activity masking of PVQ
However, the distortion function of Daala should be also added
to complete the activity masking working.

Note that PVQ QM matrix (i.e. scaler for each band of
transform block) is calculated in decoder side as exactly same
way in encoder. In Daala, this matrix is written to bitstream
and decoder does not generate it.

Activity masking can be turned off by setting below flag as 0:

Change-Id: I44bfb905cb4e0cad6aa830a4c355cd760a993ffe