Update patch set 3
Patch Set 3:

Rebased to fix the dependency on simd_check, and reinterpret_cast in the MAP macro changed to C style cast.  I think VS is in error not accepting reinterpret_cast on template function instances, but casts should make all compilers happy.

I've found one colleague with a Windows computer with VS14 and found that it now compiles with -target x86_64-win64-vs14 (disregarding build failures due to gtest errors in other tests), but it fails to compile with -target x86-win32-vs14 with alignment errors.  So it seems related to whether we're compiling for 32 bit or 64 bit.  One solution is to disable SIMD optimisations when compiling with VS for 32 bit targets, but this is a compiler limitation and not an architecture limitation, so workarounds should be investigated.  I'm not sure yet what the problem is.  Possibly, the problem is that VS for some reason does not accept a struct { __m128i, __m128i } passed as a function argument when the target is 32 bit.

Patch-set: 3
Label: Code-Review=0