Update patch set 3
Patch Set 3:

> in the end do you want to make this contingent on NDEBUG or some other form of detecting a test build?
It there is a special build mode adding a lot of tests to the code, the C implementation of intrinsics could be selected and SIMD_CHECK set accordingly, which will test for usage leading to undefined behaviour (such as shifts outside the valid range).  This will cause some unit tests which verify that SIMD optimised functions run faster than their C versions, to fail, though (these could be disabled in that mode).

I don't find NDEBUG much used (one use in decodeframe.c only?)  When is it set and for what purpose?

Patch-set: 3