Update patch set 3
Patch Set 3:

> this controls things like assert(). I'd prefer not to have debug code in a release build

I agree.  This code is purely inline and only used in the unit tests, though.

It could be useful to activate this code in the codec build if NDEBUG is set.  It would replace native intrinsics with corresponding C code along with sanity checks as a way of testing the optimised code.  On the other hand, the unit tests already test the same thing more thoroughly and systematic.

Patch-set: 3
Label: Code-Review=0