Update patch set 2
Patch Set 2:

(1 comment)

> This patch disables the SIMD optimisations for CLPF when compiling
 > with VS for 32 bit targets.  The main problem is that 32 bit VS
 > build system does not accept a struct with __m128i as a function
 > parameter, which is critical for supporting 256 bit operations. 
 > Considering that nearly every x86 CPU sold over the past 10 years
 > has supported x86-64, people who still compile for 32 bit can't be
 > much concerned about performance.

Check your chrome build on windows, if you took the default download it's likely 32-bit.

 > The speed loss of disabling SIMD
 > optimisation for CLPF is small compared to the speed loss running
 > 32 bit instead of 64 bit.  A slight performance drop seems
 > acceptable for supporting legacy systems.  Users stuck in 32 bit
 > could switch from VS to gcc to get the optimisations.

I think if you were able to switch these functions to take pointer arguments it might work. This is how we get around ABI limitations in both win32/x64 builds with more than 3/4 __m128i parameters to a function.
This might be all right for now, but the #ifdef change should be marked with a TODO for the above to given someone a hint as to what to explore.

Patch-set: 2
Label: Verified=0