Update patch set 2
Patch Set 2:

Nathan, 

Thanks for testing, please see issue 114 and 116 for some background about this commit and the other commit. The interim goal is to clear many jenkins failures and so we can get many jenkins test back, so to prevent codebase from getting even more broken than where it is now. 

And I agree with your assessment that ext-tile experiment needs more work. It is exactly the reason that I kept both issues open, with intention to fix ext-tile experiment to make it work, with or without either ans or daala_ec.  

Hope this helps,
Yaowu

Patch-set: 2
Label: Code-Review=0