Update patch set 2
Patch Set 2:

| I think if you were able to switch these functions to take pointer arguments it might work. This is how we get around ABI limitations in both win32/x64 builds with more than 3/4 __m128i parameters to a function.

The problem isn't 3/4 __m128i parameters (I think this is an issue for only one function), but that VS doesn't support a single __m128i parameter if it's inside a struct.  The struct is needed for the v256 type on non-AVX2.  I tried to work around this by clever casting or unions, but VS still complains.

The compiler will accept a pointer to these structs, but that will affect a lot of functions and fundamentally change the API, which is based on functions receiving objects as arguments and returning new objects so that the output of one function can be a parameter of another.  Forcing the user to declare the parameters and return value for every function call will make everything hard to use.

Patch-set: 2
Label: Code-Review=0