Update patch set 3
Patch Set 3:

> The compiler will accept a pointer to these structs, but that will
 > affect a lot of functions and fundamentally change the API, which
 > is based on functions receiving objects as arguments and returning
 > new objects so that the output of one function can be a parameter
 > of another.  Forcing the user to declare the parameters and return
 > value for every function call will make everything hard to use.

possibly, I haven't really had a good look. In some cases though it can be better for the reader to break out individual operations to const temporaries rather than hiding multiple operations in one line. This is a pattern we frequently use in the newer neon intrinsics in libvpx.

Patch-set: 3