Create a separate location for segmentation mask inverse
Previously, seg_mask[0] and seg_mask[1] were inverses of eachother,
but this will not be the case when the dummy mask is replaced.
Now, seg_mask[0] and seg_mask[1] will hold the masks computed
from the first and second reference frames respectively.
The inverses of these masks will be stored in inv_seg_mask.
Change-Id: Id7445e5c7dbcca347c4ea28454e95eaf5154865a