Speed up decoder when loop-restoration is enabled
With loop-restoration enabled, the decoder slows down
significantly. This patch optimizes the loop-restoration
filters, reducing this overhead by 30-40% (depending
on source and bitrate).

On bus_cif.y4m, before/after this patch, the decoder is:
200kbps:  ~200% / ~125% slower
800kbps:   ~90% /  ~61% slower
2000kbps: ~113% /  ~73% slower
with loop restoration enabled, compared to having loop
restoration disabled. The encoder is not affected much,
but is marginally faster after this patch.

This should have no effect on the video quality,
just on encode/decode speed.

Change-Id: Ide9181118ec3a63a0335338f316505b08df2d831