AV1 offline probability optimizer
This tool is a gadget for offline probability training.

A binary executable aom_entropy_optimizer will be generated in
tools/. It parses a binary file consisting of counts written in
the format of FRAME_COUNTS in entropymode.h, and computes the
optimized probability table, which will be written to a new c file
optimized_probs.c using the format in entropymode.c.

Command line:
./aom_entropy_optimizer [directory of the count file]

The input file can be either be generated from a single run by
turning on entropy_stats experiment(counts are accumulated from
frame to frame, and finally written to counts.stt), or be collected
at a larger scale, at which a python script (will be provided soon)
can be used to aggregate multiple stats output.

Optimization for initial CDFs will be also supported later.

Change-Id: I32070721aa8059439feb6b5a3a179f1001c66bb7