AV1 offline probability optimizer
This tool is a gadget for offline probability training.

A binary executable aom_entropy_optimizer will be generated. It
parses a binary file consisting of total counts written in
accordance with the format of FRAME_COUNTS in entropymode.h, and
computes the optimized probability table, which will be written to
a new c file optimized_probs.c using the format in entropymode.c.

Command line:
./aom_entropy_optimizer [directory of the count file]

The input file can be generated for a single clip if turning on
entropy_stats experiment(counts are accumulated from frame to frame,
and finally written to counts.stt). We will provide a script which
generate an file consisting of accumulated counts from multiple clips.

Change-Id: I32070721aa8059439feb6b5a3a179f1001c66bb7