Update patch set 2
Patch Set 2: Code-Review-1

I don't like this because it creates (I think) 13 macro branches just to deal with the fact that you might not want to update a CDF. This seems like a maintainability issue to me, and it will get worse if more syntax elements are coded this way. The intent of EC_ADAPT is to adapt the CDFs introduced in EC_MULTISYMBOL - if it's on it's clear that the CDFs are adapted. I would prefer an explicit non-adaptive function for the rare case you don't want to adapt the CDF even when EC_ADAPT is on.

Patch-set: 2
Reviewer: Gerrit User 5065 <5065@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-1