Correct the misleading codes in encode_inter_mb_segment()
In encode_inter_mb_segment(), when BLOCK_8X4 or BLOCK_4X8 is
passed, the nested loop inside it iterates always twice.
(For BLOCK_4X4, loop iterates only once because encode_inter_mb_segment()
is called for each of 4X4 block.)
Then, the k for 1st iteration is always zero, and the k for 2nd
iteration is always (idy * 2 + idx) with either idy == 1 or idx == 1
depending on the sb_type.

Using "+=" there could mislead readers expecting that
the # of iterations is more.
And probably using simple assignment would be more proper here.

Change-Id: I7a11255eca13403bc090ba4f0cd4785db9f0e541