Update patch set 1
Patch Set 1:

James, does this fix the VS issues?

snprintf appears to be C++11, but it's only used to format hex numbers which is trivial to write.

I'm not sure what the align issue is, but if VS is unable to align to a sizeof(), a workaround is to align to 32 always instead.

Patch-set: 1