[DO NOT SUBMIT] Palette: Don't store tokens for pixels outside image boundary.
If part of a block falls outside right and/or bottom image boundary,
then only store tokens for the part of it within the boundary.

Also, consider only the part of the block within the boundary when
calculating the number of colors in the image, deciding the base
colors for palette, RD calculation etc.

This should ideally save some bytes, but somehow BDRate gets 0.006
worse. Likely suspect is the fact that prediction is done for the whole
block even if part of it is outside image boundary.

Change-Id: Icbfbcad280e014f1bae2e1cd24edc9261c14ec25