[DO NOT SUBMIT] Palette: Ignore pixel outside boundary (Take 2)
We pad the extra indices in the color map with zeros so that same
indexes work for both prediction block and color map.

This should ideally save some bytes, but somehow BDRate gets 0.035
worse. Likely suspect is the fact that prediction is done for the whole
block even if part of it is outside image boundary.

Change-Id: I2aaef170d9317e323f6bf9693611f5960df67a4c