Replace Skip with AC/DC coded in PVQ
Instead of returning skip, av1_pvq_encode_helper and od_pvq_encode now
return ac_dc_coded. This gives more information on whether the DC part
or the AC part was skipped.

Although it is possible to obtain ac_dc_coded from the pvq_info struct,
this struct is not always used, in which case the information was lost.

This change does not impact the bitstream.

Change-Id: Ie303de915f74e8da384f822332eb1aa27f677bd3