Add a new experiment, DAALA_DIST
This commit adds the new experiment, Daala's distortion function,
which is designed to measure the activities in a block of 8x8 pixels.

Note that this experiment is expected to best working with PVQ.

The activity is measured from the overlapped 4x4 variances in the 8x8 area,
then it is used to scale the MSE distortion of 8x8 block.

Since AV1 calculates distortion down to 4x4 area, it is not possible to
directly replace the existing distortion function of AV1,
such as dist_block() and block_rd_txf(), with new one.
Hence, there has been substantial changes in order to apply
daala's 8x8 distortion function, where daala distortion function is applied
after all 4x4 tx blocks are encoded (during RDO), such as
1) intra/inter sub8x8 predictions and
2) 4x4 transform with prediction size >= 8.

To enable this experiment, add '--enable-daala-dist' with configure.

TODO: Significant tuning of parameters is required since the function has
originally came from Daala thus most parameters would not work
correctly outside Daala.
The fact that chroma distortion is added to the distortion of AV1's RDO is
also critical since Daala's distortion function is applied to luma only.

Change-Id: If35fdd3aec7efe401f351ba1c99891ad57a3d957