Update patch set 3
Patch Set 3:

The counts are aggregated in the usual way, and these are used to update the frame context binary probabilities with backward adaptation (and could be used for forward adaptation if we wanted to do that). 

In theory we could update the frame context from the adapted CDFs. But the adapted CDF probabilities represent the state at the end of a tile, not the average probabilities over the whole tile, so they may not be such a good starting point as the counts.

Patch-set: 3
Label: Code-Review=0