Update patch set 2
Patch Set 2:

The solution we use for this in Daala is to #include <math.h> from odintirn.h, and then conditionally #define M_PI if math.h does not define it.

I see the conditional #define did not get copied over when odintrin.h was ported to libaom, but that seems fixable. Are there good reasons to diverge from the Daala approach here?

Patch-set: 2
Label: Code-Review=0