Add a new experiment comp_tripred
This experiment implements the idea of the uniform tri-prediction,
which adds a 3rd inter prediction on top of the existing compound
inter bi-prediction. The final predictor for an inter-coded block is
the average of the compound bi-predictor and the third predictor,
which is indentified by a third reference frame and a third motion
vector.

Note that:
(1) Currently comp_tripred is built on top of the experiment of
ext_refs, even though it does not have to be coupled with ext_refs;
(2) The experiment of comp_tripred has not worked together with
ref_mv yet, which is on the top of the to-do list.

Compared against ext_refs, comp_tripred obtains a gain of:
lowres: Avg -1.055%; BDRate -0.946%
, where the gain is more manifest in the mid-high bitrate range.

Change-Id: I9ebcf0589e3605494a328c9bdf4aef6627378d00