Add a new experiment comp_tripred
This experiment implements the idea of uniform tri-prediction, which
adds a 3rd inter prediction on top of the existing compound inter
bi-prediction. The final predictor for a inter-coded block is the
average of the compound bi-predictor and the third predictor, which
is signaled by a third reference frame and a third motion vector.

Currently comp_tripred is built on top of the experiment of ext_refs,
even though it does not have to be coupled with ext_refs.

Change-Id: I9ebcf0589e3605494a328c9bdf4aef6627378d00