Add a new experiment comp_tripred
This experiment implements the idea of the uniform tri-prediction,
which adds a 3rd inter prediction on top of the existing single or
compound inter prediction:

(a) Compound prediction:
The final predictor for an inter-coded block is the average of the
compound bi-predictor and the third predictor, which is identified
by a third reference frame and a third motion vector;

(b) Single Prediction:
Now this new coding tool supports a new uniform compound mode that
uses one single reference frame but with two motion vectors. The final
predictor for an inter-coded block is the average of the two inter
predictors obtained through two different motion vectors from the same
reference frame.

Note that:
(1) Currently comp_tripred is built on top of the experiment of
ext_refs, even though it does not have to be coupled with ext_refs;
(2) The experiment of comp_tripred has not worked together with
ref_mv yet, which is on the top of the to-do list.

Compared against ext_refs, comp_tripred obtains a gain of:

lowres:
(a) only: Avg -1.562%; BDRate -1.385%
(a)+(b):  Avg -2.542%; BDRate -2.311%

Change-Id: I9ebcf0589e3605494a328c9bdf4aef6627378d00