Add a new experiment COMP_TRIPRED
This experiment is targeted to implement the idea of uniform
tri-prediction, which adds a 3rd inter prediction on top of the
baseline compound inter prediction. The final predictor for a inter-
coded block is the average of the compound bi-predictor and the third
predictor, which is signaled by a third reference frame and a third
motion vector.

Change-Id: I9ebcf0589e3605494a328c9bdf4aef6627378d00