Update patch set 3
Patch Set 3:

Do you have any metrics? Also, one thing to note about the new code is that the distance metric computed over (e.g.) a 32x32 square is no longer the sum of the distances computed over all the 16x16 squares. This is due to the filtering now crossing 8x8 boundaries (but not crossing the full block boundary). It's worth checking that the code does not make that assumption.

Patch-set: 3