Update patch set 8
Patch Set 8:

(4 comments)

Regarding using a single CDF, IIRC it was around 0.2% gain to use 2 CDFs, because the head should adapt faster (speed is related to CDF size). More importantly using a single CDF will be a problem for including EOB, since then you need a value N_EOB and N_NEOB for each non-zero value, and you nearly double the size of the CDF. So the head was intended to incorporate the EOB_TOKEN.

I also think we could have fewer contexts for the tail, as its probabilities are more static, which will save some storage.

Patch-set: 8
Label: Code-Review=0