Update patch set 1
Patch Set 1:

All suggestions are fine by me.  I think this proposal will make it through in some form, so it shouldn't be a problem to define TILE_BOUNDARY_TYPE outside the #if and likewise update tile_boundary_info even if CONFIG_DEBLOCKING_ACROSS_TILES is 0.  This information may be useful to have anyway, but I also suggest that tile_boundary_info should be updated to include frame borders (when tiles are disabled, which effectively means one tile), so the code at av1/common/clpf.c:218 to add the borders can be removed.

The changes can be done in your patch first and I'll then rebase and update this patch, or you could merge this patch into yours and we'll abandon this patch.

Patch-set: 1
Label: Code-Review=0