Adjust default alignment of aom_malloc() on 32-bit systems
At least when loop-restoration is enabled, we sometimes allocate
structures which need to be 16-byte aligned (eg, filter coefficients
to be loaded into SSE registers on x86). The default alignment
on 32-bit systems was 8 bytes, resulting in occasional segfaults
with loop-restoration enabled.

This patch adjusts the default alignment to 16 bytes, so that
these allocations are safe.

BUG=aomedia:345

Change-Id: Ib1b2a42d4a2fa215dcc81ea481c51271ab068a37