Add av1_convolve_2d_facade
When convolve_round is on, av1_convolve_2d_facade will be used for
interpolation rather than av1_convolve. Will remove the experiment
code of convolve_round experiment from av1_convolve in another CL.

So far we use 4-bit rounding in the intermediate stage on top of using
post rounding for compound mode after the last stage.

This will give us roughly 0.45% gain on lowres and 0.39% on midres
Altogether, is 1.15% on lowres and 0.74% on midres.

Change-Id: I6332e1d888e28a3b3ddc29711817d66e52cb5cdf