Update patch set 16
Patch Set 16:

Hey folks, I realize we are already half way down this path, but it strikes me we shouldn't be doing this rework on the live ec_multisymbol experiment. Maybe this should be done in a separate experiment and merged into ec_multisymbol when it is stable and has acwy numbers.

Patch-set: 16