Alternate DAALA_DIST scaling for keyframes
The current curve matches inter at cq-levels 20, 32, 43 and 55.
For intra, it uses a quadratic best fit anchored to inter.
This matches the first point tightly but diverges thereafter.

We can fit better for both inter and intra with separate curves.
Use Q17 quadratics to ease transferring from notebook to code.

Change-Id: I6df360d6a17f6caa5b150a46aa73982af34aaec7