Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I wonder if you could prune all but the googletest directory to pick up just that CMakeLists.txt. The top-level CMakeLists.txt ends up including only gmock by default

Besides pulling in the cmake support one of the points of doing things this way is to stop messing around with the third_party sources and just pull in what's upstream. Ideally we won't even need googletest and libwebm in tree if things work out with cmake. Googletest is pretty tiny, can we just skip the mucking with third party code?

Patch-set: 3
Label: Code-Review=0