improving palette throughput
* code the palette color index using 45 wavefront
* interleave the coeff and palette color index in
  transform block level
* the above change does not change code efficiency

Details: 
The 45 wavefront scan allows to compute the ctx of the 
diagonal samples' indices  at the same time. 
Interleaving palette indices and palette residual on a transform block basis 
means that the entropy decoding and further processing of the palette residual 
is not delayed by the entropy decoding of all the 
color indices of the palette encoded block.

Change-Id: Ie9f576002a9a68394b99c23b01e9730df06df070