WIP: Adjust global motion rate penalty
Currently, global-motion applies a rate cost penalty per block
until we have "enough" blocks to justify keeping the global motion
model. But the decision of whether to keep the model is actually
made based on the number of pixels using global motion, not the
number of blocks.

This patch brings the rate penalty in line with the later
decision-making process.

Change-Id: I0c11b1530b1d9032c4aeda9185c54f3d45e3bb85