handle_inter_mode: add arguments struct
Adds the arguments struct HandleInterModeArgs to hold arguments that
are conditional on compiled features. This means that there are no
longer #if's in the function's argument list.

Some of the array pointers that were optional arguments have been
made array members in the new struct, but not all. This was due to the
function being called with either references to arrays that are
maintained between different modes OR with references to "dummy"
arrays initialized to zero. The arrays that are always used are now
members of the HandleInterModeArgs struct.

Change-Id: I3076fd53c3cddf5a6d14bbe7d23a889465ed716d