Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> I still get the empty file warnings for the .asm includes

Meant to ask about that: the asm includes appear to need to be built as source files in x86_64, but are only includes in x86 (well, that appears true for x86_abi_support, at least). Do I need to pull it out of the sources list when x86 is the target?

Anyway, if you see this here, great: otherwise I'll come back around to it.

Patch-set: 7
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>