Remove ATTRIBUTE_PACKED from InterpFilter which broke something
The ATTRIBUTE_PACKED caused something to break in MB_MODE_INFO.
boundary_info got corrupted and moving InterpFilter around in
MB_MODE_INFO could cause a segfault or the problems to go away (for
instance by moving InterpFilter to the end of the struct).  Perhaps
something in MB_MODE_INFO is sensitive to alignment.  This should be
investigated further and get fixed or documented, but for now remove
the attribute.

Change-Id: Ic4326ab98fe784d9b7d2b47d8f086b8429cedee9