Update patch set 1
Patch Set 1:

So one issue is that if the optional arguments struct isn't used, you'll get a compiler warning. I don't want to put #if's around it since not having them is the point of this patch.

Two possible solutions I can think of: use the "(void)object;" instruction like various places in the code base, or add one of the arguments that always gets used to the struct so that the struct itself is always used.

Patch-set: 1
Reviewer: Gerrit User 6465 <6465@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>