Chroma from Luma (CfL) for Intra in PVQ
Uses the reconstructed Luma pixel as an intra prediction for the Chroma
pixels. In this first version, CfL is always on, which requires the
UV_DC_PRED_ONLY experiment so that Chroma intra prediction is not
signaled in the bitstream.

This preliminary version of CfL only supports 4:2:0 and does not perform CfL
during RDO. PVQ's gain prediction is used (which will be disabled in a
subsequent version). CfL flipping is also not implemented yet.

The CFL_CTX struct is introduced, it stores a pixel buffer per thread group.
CFL_CTX will replace PVQ's cfl_ctx struct in a subsequent version.

When 4:2:0 is used, it is possible to have Chromatic overrun (i.e. when
the surface covered by Chroma is bigger than the surface covered by
luma). When this happens we pad by copying the last available luma pixel
value.

subset-1:
NO_Pix_CfL@2017-02-22T16:03:35.757Z -> Pix_CfL@2017-02-22T16:04:12.652Z
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
18.8998 | 11.5547 | 11.7256 |  19.9913 | 19.0182 | 19.0427 |    18.4401

Change-Id: Iac8c755c37119c818d60ab33592517d6d9389fa8